// packages/utils/src/form-assosiated/types.ts
export type FormAssosiatedProps = {
  customMessage: string
  disabled: boolean
  initValidation: boolean
  inputValidation: boolean
  max: number | undefined
  min: number | undefined
  name: string
  readonly: boolean
  reporterNode: HTMLElement | undefined
  required: boolean
  step: number | undefined
  value: string
  willValidate: boolean
}